Move Camera into the settings module
