+++ stellarphot/photometry/tests/test_profiles.py
@@ -3,9 +3,9 @@ import pytest
 from astropy.table import Table
 from photutils.datasets import make_gaussian_sources_image, make_noise_image
 
-from stellarphot import Camera
 from stellarphot.photometry import CenterAndProfile, find_center
-from stellarphot.tests.test_core import TEST_CAMERA_VALUES
+from stellarphot.settings import Camera
+from stellarphot.settings.tests.test_models import TEST_CAMERA_VALUES
 
 # Make a few round stars
 STARS = Table(
